Se ham colocado correctamente las clases y se han añadido los test de cada clase
