--- Test/es/uah/matcomp/mp/e1/ejerciciosclases/AccountTest.java
@@ -1,6 +1,0 @@
-package es.uah.matcomp.mp.e1.ejerciciosclases;
-
-
-public class AccountTest {
-
-    }
