+++ Test/es/uah/matcomp/mp/e1/ejerciciosclases/Practica1/AccountTest.java
@@ -1,0 +1,36 @@
+package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class AccountTest {
+
+    @Test
+    void getID() {
+    }
+
+    @Test
+    void getName() {
+    }
+
+    @Test
+    void getBalance() {
+    }
+
+    @Test
+    void credit() {
+    }
+
+    @Test
+    void debit() {
+    }
+
+    @Test
+    void transferTo() {
+    }
+
+    @Test
+    void testToString() {
+    }
+}
