+++ Test/es/uah/matcomp/mp/e1/ejerciciosclases/Practica1/BallTest.java
@@ -1,0 +1,64 @@
+package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class BallTest {
+
+    @Test
+    void getX() {
+    }
+
+    @Test
+    void getY() {
+    }
+
+    @Test
+    void setX() {
+    }
+
+    @Test
+    void setY() {
+    }
+
+    @Test
+    void getRadius() {
+    }
+
+    @Test
+    void setRadius() {
+    }
+
+    @Test
+    void getXDelta() {
+    }
+
+    @Test
+    void setXDelta() {
+    }
+
+    @Test
+    void getYDelta() {
+    }
+
+    @Test
+    void setYDelta() {
+    }
+
+    @Test
+    void move() {
+    }
+
+    @Test
+    void reflectHorizontal() {
+    }
+
+    @Test
+    void reflectVertical() {
+    }
+
+    @Test
+    void testToString() {
+    }
+}
