+++ Test/es/uah/matcomp/mp/e1/ejerciciosclases/Practica1/CircleTest.java
@@ -1,0 +1,28 @@
+package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class CircleTest {
+
+    @Test
+    void getRadius() {
+    }
+
+    @Test
+    void getArea() {
+    }
+
+    @Test
+    void getCircumference() {
+    }
+
+    @Test
+    void testToString() {
+    }
+
+    @Test
+    void setRadius() {
+    }
+}
