+++ Test/es/uah/matcomp/mp/e1/ejerciciosclases/Practica1/DateTest.java
@@ -1,0 +1,40 @@
+package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class DateTest {
+
+    @Test
+    void getDay() {
+    }
+
+    @Test
+    void getMonth() {
+    }
+
+    @Test
+    void getYear() {
+    }
+
+    @Test
+    void setDay() {
+    }
+
+    @Test
+    void setMonth() {
+    }
+
+    @Test
+    void setYear() {
+    }
+
+    @Test
+    void setDate() {
+    }
+
+    @Test
+    void testToString() {
+    }
+}
