+++ Test/es/uah/matcomp/mp/e1/ejerciciosclases/Practica1/EmployeeTest.java
@@ -1,0 +1,44 @@
+package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class EmployeeTest {
+
+    @Test
+    void getId() {
+    }
+
+    @Test
+    void getFirstName() {
+    }
+
+    @Test
+    void getLastName() {
+    }
+
+    @Test
+    void getName() {
+    }
+
+    @Test
+    void getSalary() {
+    }
+
+    @Test
+    void getAnnualSalary() {
+    }
+
+    @Test
+    void setSalary() {
+    }
+
+    @Test
+    void raiseSalary() {
+    }
+
+    @Test
+    void testToString() {
+    }
+}
