+++ Test/es/uah/matcomp/mp/e1/ejerciciosclases/Practica1/InvoiceItemTest.java
@@ -1,0 +1,40 @@
+package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class InvoiceItemTest {
+
+    @Test
+    void getId() {
+    }
+
+    @Test
+    void getDesc() {
+    }
+
+    @Test
+    void getQty() {
+    }
+
+    @Test
+    void getUnitPrice() {
+    }
+
+    @Test
+    void getTotal() {
+    }
+
+    @Test
+    void setQty() {
+    }
+
+    @Test
+    void setUnitPrice() {
+    }
+
+    @Test
+    void testToString() {
+    }
+}
