+++ Test/es/uah/matcomp/mp/e1/ejerciciosclases/Practica1/RectangleTest.java
@@ -1,0 +1,36 @@
+package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class RectangleTest {
+
+    @Test
+    void getLength() {
+    }
+
+    @Test
+    void getWidth() {
+    }
+
+    @Test
+    void setLength() {
+    }
+
+    @Test
+    void setWidth() {
+    }
+
+    @Test
+    void getArea() {
+    }
+
+    @Test
+    void getPerimeter() {
+    }
+
+    @Test
+    void testToString() {
+    }
+}
