+++ Test/es/uah/matcomp/mp/e1/ejerciciosclases/Practica1/TimeTest.java
@@ -1,0 +1,48 @@
+package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class TimeTest {
+
+    @Test
+    void getHour() {
+    }
+
+    @Test
+    void getMinute() {
+    }
+
+    @Test
+    void getSecond() {
+    }
+
+    @Test
+    void setHour() {
+    }
+
+    @Test
+    void setMinute() {
+    }
+
+    @Test
+    void setSecond() {
+    }
+
+    @Test
+    void setTime() {
+    }
+
+    @Test
+    void testToString() {
+    }
+
+    @Test
+    void nextSecond() {
+    }
+
+    @Test
+    void previousSecond() {
+    }
+}
