--- src/es/uah/matcomp/mp/e1/ejerciciosclases/Main.java
@@ -1,5 +1,0 @@
-package es.uah.matcomp.mp.e1.ejerciciosclases;
-
-public class MainAccount {
-
-}
