+++ src/es/uah/matcomp/mp/e1/ejerciciosclases/Practica1/Time.java
@@ -1,4 +1,4 @@
-package es.uah.matcomp.mp.e1.ejerciciosclases;
+package es.uah.matcomp.mp.e1.ejerciciosclases.Practica1;
 
 public class Time {
     private int hour;
